+++ lib/main.dart
@@ -17,8 +17,7 @@ class MyApp extends StatelessWidget {
         colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
         useMaterial3: true,
       ),
-      home:  DiaryWorldPage(),
+      home: DiaryWorldPage(),
     );
   }
 }
-